+++ CHANGELOG.md
@@ -5,6 +5,7 @@ https://keepachangelog.com/en/1.0.0/
 - `StringConcatenatedInLoop`: Now also catches interpolated strings
 - `StringPlaceholdersInWrongOrder`: Retains existing whitespace when applying the code fix
 - `LinqTraversalBeforeFilter`: Implemented a code fix to swap the traversal and filter calls
+- `ComparingStringsWithoutStringComparison`: Supports null-conditional access, e.g. `myString?.ToLower() == otherString?.ToLower()`
 
 ## [1.29.0] - 2026-01-22
 - Fixed an ignored test `MultipleOrderByCalls_MultipleOrderByChained`